Move `MoreCollections` to `io.spine.base.util`
